Fix bug in deep datatype https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki#serialization-format ref #339

+++ skycoin-api/tools/bip32.h
@@ -42,7 +42,7 @@ typedef struct {
 } curve_info;
 
 typedef struct {
-    uint32_t depth;
+    uint8_t depth;
     uint32_t child_num;
     uint8_t chain_code[32];
